비동기 action 테스트 작성 시도
