+++ src/App.jsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import RegionListContainer from './RegionListContainer';
 import CategoryListContainer from './CategoryListContainer';
-import { loadCategories, loadRegions } from './store/actions';
 import RestaurantListContainer from './RestaurantListContainer';
+import { loadCategories, loadRegions } from './store/asyncActions';
 
 export default function App() {
   const dispatch = useDispatch();
