+++ src/store/actions.js
@@ -1,6 +1,3 @@
-import { fetchCategories, fetchRegions, fetchRestaurants } from '../services';
-import { isNull } from '../lib';
-
 export const CHANGE_REGION = 'changeRegion';
 export const CHANGE_CATEGORY = 'changeCategory';
 export const SET_REGIONS = 'setRegions';
@@ -49,38 +46,5 @@ export function setRestaurants(restaurants) {
     payload: {
       restaurants,
     },
-  };
-}
-
-export function loadRegions() {
-  return async (dispatch) => {
-    const regions = await fetchRegions();
-
-    dispatch(setRegions(regions));
-  };
-}
-
-export function loadCategories() {
-  return async (dispatch) => {
-    const categories = await fetchCategories();
-
-    dispatch(setCategories(categories));
-  };
-}
-
-export function loadRestaurants() {
-  return async (dispatch, getState) => {
-    const { selected: { regionName, categoryId } } = getState();
-
-    if (isNull(regionName) || isNull(categoryId)) {
-      return;
-    }
-
-    const restaurants = await fetchRestaurants({
-      regionName,
-      categoryId,
-    });
-
-    dispatch(setRestaurants(restaurants));
   };
 }
